+++ packages/skeleton-react/src/components/accordion/modules/anatomy.ts
@@ -1,6 +1,5 @@
 import Item from '../anatomy/item';
 import ItemContent from '../anatomy/item-content';
-import ItemHeading from '../anatomy/item-heading';
 import ItemIndicator from '../anatomy/item-indicator';
 import ItemTrigger from '../anatomy/item-trigger';
 import Root from '../anatomy/root';
@@ -11,7 +10,6 @@ export const Accordion = Object.assign(Root, {
 	Provider: RootProvider,
 	Context: RootContext,
 	Item: Item,
-	ItemHeading: ItemHeading,
 	ItemTrigger: ItemTrigger,
 	ItemIndicator: ItemIndicator,
 	ItemContent: ItemContent,