+++ packages/skeleton-react/src/components/date-picker/modules/root-context.ts
@@ -1,0 +1,4 @@
+import type { useDatePicker } from './provider';
+import { createContext } from '@/internal/create-context';
+
+export const RootContext = createContext<ReturnType<typeof useDatePicker>>();