+++ sites/skeleton.dev/src/examples/components/accordion/react/controlled.tsx
@@ -7,23 +7,23 @@ export default function Controlled() {
 	return (
 		<Accordion value={value} onValueChange={(e) => setValue(e.value)}>
 			<Accordion.Item value="item-1">
-				<Accordion.ItemHeading>
+				<h3>
 					<Accordion.ItemTrigger>Item 1</Accordion.ItemTrigger>
-				</Accordion.ItemHeading>
+				</h3>
 				<Accordion.ItemContent>Content for Item 1</Accordion.ItemContent>
 			</Accordion.Item>
 			<hr className="hr" />
 			<Accordion.Item value="item-2">
-				<Accordion.ItemHeading>
+				<h3>
 					<Accordion.ItemTrigger>Item 2</Accordion.ItemTrigger>
-				</Accordion.ItemHeading>
+				</h3>
 				<Accordion.ItemContent>Content for Item 2</Accordion.ItemContent>
 			</Accordion.Item>
 			<hr className="hr" />
 			<Accordion.Item value="item-3">
-				<Accordion.ItemHeading>
+				<h3>
 					<Accordion.ItemTrigger>Item 3</Accordion.ItemTrigger>
-				</Accordion.ItemHeading>
+				</h3>
 				<Accordion.ItemContent>Content for Item 3</Accordion.ItemContent>
 			</Accordion.Item>
 		</Accordion>