+++ sites/skeleton.dev/src/examples/components/accordion/svelte/columns.svelte
@@ -8,42 +8,42 @@
 
 <Accordion>
 	<Accordion.Item value="club">
-		<Accordion.ItemHeading>
+		<h3>
 			<Accordion.ItemTrigger class="grid grid-cols-[1fr_auto]">
 				<span>Club</span>
 				<IconClub />
 			</Accordion.ItemTrigger>
-		</Accordion.ItemHeading>
+		</h3>
 		<Accordion.ItemContent>Content for Item 1</Accordion.ItemContent>
 	</Accordion.Item>
 	<hr class="hr" />
 	<Accordion.Item value="diamond">
-		<Accordion.ItemHeading>
+		<h3>
 			<Accordion.ItemTrigger class="grid grid-cols-[1fr_auto]">
 				<span>Diamond</span>
 				<IconDiamond />
 			</Accordion.ItemTrigger>
-		</Accordion.ItemHeading>
+		</h3>
 		<Accordion.ItemContent>Content for Item 2</Accordion.ItemContent>
 	</Accordion.Item>
 	<hr class="hr" />
 	<Accordion.Item value="heart">
-		<Accordion.ItemHeading>
+		<h3>
 			<Accordion.ItemTrigger class="grid grid-cols-[1fr_auto]">
 				<span>Heart</span>
 				<IconHeart />
 			</Accordion.ItemTrigger>
-		</Accordion.ItemHeading>
+		</h3>
 		<Accordion.ItemContent>Content for Item 3</Accordion.ItemContent>
 	</Accordion.Item>
 	<hr class="hr" />
 	<Accordion.Item value="spade">
-		<Accordion.ItemHeading>
+		<h3>
 			<Accordion.ItemTrigger class="grid grid-cols-[1fr_auto]">
 				<span>Spade</span>
 				<IconSpade />
 			</Accordion.ItemTrigger>
-		</Accordion.ItemHeading>
+		</h3>
 		<Accordion.ItemContent>Content for Item 4</Accordion.ItemContent>
 	</Accordion.Item>
 </Accordion>